Implement Map Based Services. Closes #5.
As all the Service implementations will be rather similar, let's create an AbstractMapService based on generics too.

Then all the actual implementations can extend from this abstract service and only provide the concrete Type and ID values and just call the methods of the super class they are extending from.