Remove pets from the Owner Constructor used for the Builder pattern as this leads to a chicken and egg problem.
If you don't set pets via the builder this value is null and the DataInitializer will crash when try to add a pet via:

mike.getPets().add(mikesPet);

However, you can't properly set the pets via the builder as a pet has a reference to its owner which creates a circular reference.
What you could do is do .pets(new HashSet()) in the builder, but this seems silly.